Don't test variable-length array by default
Variable-length arrays (VLA) are a feature of C99, that was made
optional in C11. A compliant C11 compiler defines __STDC_NO_VLA__ to 1
if it doesn't support VLAs.

MSVC doesn't implement VLAs in C99 (non-compliant) but defines
__STDC_NO_VLA__ in C11 mode. If we don't guard the VLA in
_AC_C_C99_TEST_MAIN, as the test is included in
_AC_C_C11_TEST_PROGRAM, MSVC won't be considered as a compiler
supporting C11 (which, to some extent, it does).

Fix this by guarding the C99 code using VLA with the C11 macro
__STDC_NO_VLA__.